Data on model metalearners experiments

metalearners arose from a set of experiments intended to benchmark the time-to-fit and predictive performance of a number of alternative model stacking methods (or "meta-learners") for use in the stacks package. Each row in the data represents a unique combination of task, proposed meta-learner (as recipe + specification), and performance metric. The meta-learner minimal_glmnet is the existing meta-learner, and the remaining entries represent alternative meta-learners to be compared to minimal_glmnet.

Usage

metalearners

Format

metalearners

A data frame with 217 rows and 8 columns:

task

Factor. A descriptor of a modeling task, i.e. a dataset.

meta_learner

Factor. An ID for the proposed meta-learner. A proposed meta-learner is made by pairing a recipe with a model specification.

recipe

Factor. The recipe used to preprocess the data stack. See the meta_learners folder in the linked repository for implementation.

specification

Factor. The model specification used to model the data stack. See the meta_learners folder in the linked repository for implementation.

elapsed

Numeric. The elapsed time to tune the meta-learner against resamples, in seconds. In "real" time rather than CPU time.

metric

Character. The type of performance metric, as returned by the yardstick package.

estimate

Character. The performance estimate, as returned by the yardstick package.

estimator

Character. The performance estimator, as returned by the yardstick package.

Source

https://github.com/simonpcouch/metalearner

Details

Every meta_learner was evaluated against every task. That said, a number of meta-learners failed to train on a given task, usually due to memory exhaustion, and those failed experiments are not represented in the data.

Each experiment was carried out on a 24-core machine with 64GB memory running Linux Ubuntu 24.04. Computations were distributed using forking with 5 cores (except for tasks caret and nhl, which used 2 cores.)

Examples

str(metalearners)
#> tibble [217 × 8] (S3: tbl_df/tbl/data.frame)
#>  $ task         : Factor w/ 8 levels "barley","caret",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ meta_learner : Factor w/ 11 levels "minimal_glmnet",..: 1 1 2 2 3 3 4 4 5 5 ...
#>  $ recipe       : Factor w/ 4 levels "C+S, PCA, C+S",..: 3 3 3 3 3 3 2 2 2 2 ...
#>  $ specification: Factor w/ 6 levels "Bagged Tree (rpart)",..: 4 4 2 2 3 3 1 1 5 5 ...
#>  $ elapsed      : num [1:217] 0.628 0.628 194.429 194.429 263.355 ...
#>  $ metric       : chr [1:217] "rsq" "rmse" "rsq" "rmse" ...
#>  $ estimate     : num [1:217] 0.977 2.288 0.966 3.134 0.973 ...
#>  $ estimator    : chr [1:217] "standard" "standard" "standard" "standard" ...

metalearners
#> # A tibble: 217 × 8
#>    task   meta_learner   recipe  specification elapsed metric estimate estimator
#>    <fct>  <fct>          <fct>   <fct>           <dbl> <chr>     <dbl> <chr>    
#>  1 barley minimal_glmnet Minimal GLM (glmnet)    0.628 rsq       0.977 standard 
#>  2 barley minimal_glmnet Minimal GLM (glmnet)    0.628 rmse      2.29  standard 
#>  3 barley minimal_lgb    Minimal Boosted Tree… 194.    rsq       0.966 standard 
#>  4 barley minimal_lgb    Minimal Boosted Tree… 194.    rmse      3.13  standard 
#>  5 barley minimal_xgb    Minimal Boosted Tree… 263.    rsq       0.973 standard 
#>  6 barley minimal_xgb    Minimal Boosted Tree… 263.    rmse      2.53  standard 
#>  7 barley normalize_bt   Center… Bagged Tree …  25.7   rsq       0.970 standard 
#>  8 barley normalize_bt   Center… Bagged Tree …  25.7   rmse      2.63  standard 
#>  9 barley normalize_nn   Center… Neural Netwo… 407.    rsq       0.859 standard 
#> 10 barley normalize_nn   Center… Neural Netwo… 407.    rmse      5.75  standard 
#> # ℹ 207 more rows